#23-Make GetCurrentDateFile for using function more conveniently
